refactor(cli): extract the Tuist.swift scanner into its own module
The cache-provider resolver inlined a hand-rolled scanner that walks a
Tuist.swift manifest to read the project handle and cache URL. Moving it
to a focused submodule with a single entry point removes about 180 lines
from the resolver and isolates the scanning logic behind one function.
