+++ crates/once-cli/src/cache_provider.rs
@@ -11,6 +11,8 @@ use once_frontend::{
 };
 use serde::Deserialize;
 
+mod tuist_swift_config;
+
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) enum ResolvedCacheProviderConfig {
     Local,
@@ -433,192 +435,8 @@ fn load_tuist_toml_config(workspace: &Path) -> Option<TuistWorkspaceConfig> {
 fn load_tuist_swift_config(workspace: &Path) -> Option<TuistWorkspaceConfig> {
     let path = workspace.join("Tuist.swift");
     let src = std::fs::read_to_string(path).ok()?;
-    let body = swift_tuist_constructor_body(&src)?;
-    Some(TuistWorkspaceConfig {
-        full_handle: swift_string_argument(body, "fullHandle")?,
-        url: swift_string_argument(body, "url"),
-    })
-}
-
-fn swift_tuist_constructor_body(src: &str) -> Option<&str> {
-    let mut cursor = 0;
-    while cursor < src.len() {
-        if let Some(next) = skip_swift_comment_or_string(src, cursor) {
-            cursor = next;
-            continue;
-        }
-        if starts_swift_identifier(src, cursor, "Tuist") {
-            let open = skip_swift_space_and_comments(src, cursor + "Tuist".len())?;
-            if src[open..].starts_with('(') {
-                return swift_parenthesized_body(src, open);
-            }
-        }
-        cursor += src[cursor..].chars().next()?.len_utf8();
-    }
-    None
-}
-
-fn swift_parenthesized_body(src: &str, open: usize) -> Option<&str> {
-    let mut cursor = open + 1;
-    let mut depth = 1;
-    while cursor < src.len() {
-        if let Some(next) = skip_swift_comment_or_string(src, cursor) {
-            cursor = next;
-            continue;
-        }
-        let ch = src[cursor..].chars().next()?;
-        if ch == '(' {
-            depth += 1;
-        } else if ch == ')' {
-            depth -= 1;
-            if depth == 0 {
-                return Some(&src[open + 1..cursor]);
-            }
-        }
-        cursor += ch.len_utf8();
-    }
-    None
-}
-
-fn swift_string_argument(src: &str, label: &str) -> Option<String> {
-    let mut cursor = 0;
-    let mut depth = 0usize;
-    while cursor < src.len() {
-        if let Some(next) = skip_swift_comment_or_string(src, cursor) {
-            cursor = next;
-            continue;
-        }
-        let ch = src[cursor..].chars().next()?;
-        if matches!(ch, '(' | '[' | '{') {
-            depth += 1;
-            cursor += ch.len_utf8();
-            continue;
-        }
-        if matches!(ch, ')' | ']' | '}') {
-            depth = depth.saturating_sub(1);
-            cursor += ch.len_utf8();
-            continue;
-        }
-        if depth == 0 && starts_swift_identifier(src, cursor, label) {
-            let colon = skip_swift_space_and_comments(src, cursor + label.len())?;
-            if src[colon..].starts_with(':') {
-                return leading_swift_string(&src[colon + 1..]);
-            }
-        }
-        cursor += ch.len_utf8();
-    }
-    None
-}
-
-fn starts_swift_identifier(src: &str, index: usize, identifier: &str) -> bool {
-    if !src[index..].starts_with(identifier) {
-        return false;
-    }
-    let before = src[..index].chars().next_back();
-    let after = src[index + identifier.len()..].chars().next();
-    !before.is_some_and(is_swift_identifier_char) && !after.is_some_and(is_swift_identifier_char)
-}
-
-fn is_swift_identifier_char(ch: char) -> bool {
-    ch == '_' || ch.is_ascii_alphanumeric()
-}
-
-fn skip_swift_space_and_comments(src: &str, mut cursor: usize) -> Option<usize> {
-    loop {
-        while cursor < src.len() {
-            let ch = src[cursor..].chars().next()?;
-            if !ch.is_whitespace() {
-                break;
-            }
-            cursor += ch.len_utf8();
-        }
-        if let Some(next) = skip_swift_comment(src, cursor) {
-            cursor = next;
-        } else {
-            return Some(cursor);
-        }
-    }
-}
-
-fn skip_swift_comment_or_string(src: &str, cursor: usize) -> Option<usize> {
-    skip_swift_comment(src, cursor).or_else(|| skip_swift_string(src, cursor))
-}
-
-fn skip_swift_comment(src: &str, cursor: usize) -> Option<usize> {
-    if src[cursor..].starts_with("//") {
-        return Some(
-            src[cursor..]
-                .find('\n')
-                .map_or(src.len(), |offset| cursor + offset + 1),
-        );
-    }
-    if !src[cursor..].starts_with("/*") {
-        return None;
-    }
-    let mut depth = 1usize;
-    let mut current = cursor + 2;
-    while current < src.len() {
-        if src[current..].starts_with("/*") {
-            depth += 1;
-            current += 2;
-        } else if src[current..].starts_with("*/") {
-            depth -= 1;
-            current += 2;
-            if depth == 0 {
-                return Some(current);
-            }
-        } else {
-            current += src[current..].chars().next()?.len_utf8();
-        }
-    }
-    Some(src.len())
-}
-
-fn skip_swift_string(src: &str, cursor: usize) -> Option<usize> {
-    if !src[cursor..].starts_with('"') {
-        return None;
-    }
-    if src[cursor..].starts_with("\"\"\"") {
-        return Some(
-            src[cursor + 3..]
-                .find("\"\"\"")
-                .map_or(src.len(), |offset| cursor + 3 + offset + 3),
-        );
-    }
-    let mut current = cursor + 1;
-    let mut escaped = false;
-    while current < src.len() {
-        let ch = src[current..].chars().next()?;
-        current += ch.len_utf8();
-        if escaped {
-            escaped = false;
-        } else if ch == '\\' {
-            escaped = true;
-        } else if ch == '"' {
-            return Some(current);
-        }
-    }
-    Some(src.len())
-}
-
-fn leading_swift_string(src: &str) -> Option<String> {
-    let trimmed = src.trim_start();
-    let body = trimmed.strip_prefix('"')?;
-    let mut value = String::new();
-    let mut escaped = false;
-    for ch in body.chars() {
-        if escaped {
-            value.push(ch);
-            escaped = false;
-        } else if ch == '\\' {
-            escaped = true;
-        } else if ch == '"' {
-            return non_empty(Some(value));
-        } else {
-            value.push(ch);
-        }
-    }
-    None
+    let (full_handle, url) = tuist_swift_config::parse_tuist_config(&src)?;
+    Some(TuistWorkspaceConfig { full_handle, url })
 }
 
 fn split_full_handle(full_handle: &str) -> Option<(String, String)> {
